UsersProject/requirements.txt
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.8.0
APScheduler==3.11.0
arrow==1.3.0
asgiref==3.8.1
bcrypt==4.2.1
blessed==1.20.0
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
dj-database-url==2.3.0
Django==4.2.18
django-cors-headers==4.6.0
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.15.2
exceptiongroup==1.2.2
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
jinxed==1.3.0
MarkupSafe==3.0.2
numpy==2.0.2
pandas==2.2.3
paramiko==3.5.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
PyNaCl==1.5.0
pypdf==5.2.0
pypdf2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
redis==3.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
starlette==0.45.3
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uvicorn==0.34.0
wcwidth==0.2.13
werkzeug==3.1.3
